requirements.txt
bcrypt==4.2.0
charset-normalizer==2.0.12
cryptography==40.0.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.2.0
expiringdict==1.2.2
fastapi==0.109.2
httptools==0.6.1
httpx==0.27.2
importlib-metadata==5.0.0
ipython==8.6.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
multidict==6.1.0
numpy==1.21.6
openai==0.28.0
openapi-codec==1.3.2
openpyxl==3.1.5
pandas==1.3.5
passlib==1.7.4
prompt-toolkit==3.0.32
pydantic==1.10.7
pyparsing==3.0.9
pyrsistent==0.19.3
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.4
python-multipart==0.0.9
pytz==2019.3
redis==4.5.4
requests==2.28.0
requests-toolbelt==0.10.1
SQLAlchemy==2.0.35
sqlparse==0.4.4
tenacity==8.1.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn[standard]
xmljson==0.2.0
yarl==1.11.1
zipp==3.10.0
sqlalchemy
